+++ src/trackers/index.ts
@@ -21,7 +21,13 @@ export type { GameTrackerState } from '#trackers/game-tracker';
 export type { PieceTrackerState } from '#trackers/piece-tracker';
 export type { TileTrackerState } from '#trackers/tile/tile-tracker';
 export type { HeatmapData, HeatmapFn } from '#trackers/heatmap-types';
-export type { Action, CaptureAction, MoveAction, MoveCoords, PromoteAction } from '#types/actions';
+export type {
+    Action,
+    CaptureAction,
+    MoveAction,
+    MoveCoords,
+    PromotionAction,
+} from '#types/actions';
 export type { TrackerInstance } from '#types/tracker';
 export type {
     PlayerSquareStats,